requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
sentence-transformers
groq
python-dotenv
duckduckgo-search